refactor(stake-meta): extract output types into lightweight crate
